requirements.txt
streamlit
pandas
plotly
reportlab
Pillow
openpyxl
io